fix error <= 2 arg

+++ includes/push_swap.h
@@ -35,9 +35,10 @@ typedef enum e_operation
 
 typedef struct s_list_stack
 {
-	int					content;
+	long					content;
 	struct s_list_stack	*next;
 }				t_list_s;
+
 typedef struct s_stack
 {
 	t_list_s	*stack_a;
